Refactored a few files in /lib (#1850)
* chore: Marked isValidLocaleFormat as deprecated

* refactor: Removed libs/utils/is-engine

* refactor: Read ordinals.json and plurals.json from cldr-core in one place

* chore: Added changeset

---------
