+++ .changeset/rare-forks-develop.md
@@ -1,0 +1,5 @@
+---
+"ember-intl": patch
+---
+
+Refactored a few files in /lib
