+++ packages/ember-intl/lib/broccoli/translation-reducer/utils/is-known-language.js
@@ -1,9 +1,4 @@
-const ordinals = require('cldr-core/supplemental/ordinals.json').supplemental[
-  'plurals-type-ordinal'
-];
-const plurals = require('cldr-core/supplemental/plurals.json').supplemental[
-  'plurals-type-cardinal'
-];
+const { ordinals, plurals } = require('../../../utils/cldr-core');
 
 module.exports = function isKnownLanguage(language) {
   return language in ordinals || language in plurals;
