--- packages/ember-intl/lib/utils/is-engine.js
@@ -1,9 +1,0 @@
-function isEngine(addon) {
-  return (
-    addon.pkg &&
-    addon.pkg.keywords &&
-    addon.pkg.keywords.includes('ember-engine')
-  );
-}
-
-module.exports = isEngine;
