+++ packages/ember-intl/lib/utils/is-valid-locale-format.js
@@ -1,3 +1,8 @@
+/**
+ * @deprecated Will be removed in v7.
+ *
+ * Do not use it.
+ */
 function isValidLocaleFormat(locale) {
   if (!/^[a-zA-Z0-9-]+$/.test(locale)) {
     return false;
